+++ pages/models/building_operation/chilling.py
@@ -186,7 +186,9 @@ class CoolingSystemChiller(models.Model):
     baseline_leakage_factor_percent = models.PositiveIntegerField(
         default=2, verbose_name=_("Baseline Leakage Factor (%)")
     )
-    total_energy_consumption_kwh_per_year = models.PositiveIntegerField(
+    total_energy_consumption_kwh_per_year = models.DecimalField(
+        max_digits=12,
+        decimal_places=3,
         null=True,
         blank=True,
         verbose_name=_(
